more cleanup

+++ examples/jsm/nodes/accessors/BatchNode.js
@@ -45,11 +45,11 @@ class BatchNode extends Node {
 			const size = textureSize( textureLoad( this.batchMesh._indirectTexture ), 0 );
 			const x = int( id ).remainder( int( size ) );
 			const y = int( id ).div( int( size ) );
-			return textureLoad( this.batchMesh._indirectTexture, ivec2( x, y ), null, 'uvec4' ).x.toFloat();
+			return textureLoad( this.batchMesh._indirectTexture, ivec2( x, y ), null, 'uvec4' ).x;
 
 		} ).setLayout( {
 			name: 'getIndirectIndex',
-			type: 'float',
+			type: 'uint',
 			inputs: [
 				{ name: 'id', type: 'int' }
 			]
